// frontend/src/pages/Blocks/types.ts
export type BlockType =
  | "Link"
  | "Email"
  | "Address"
  | "Phone"
  | "Facebook"
  | "Twitter"
  | "Instagram"
  | "Youtube"
  | "Whatsapp"
  | "Tiktok"
  | "Telegram"
  | "Spotify"
  | "Pinterest"
  | "Linkedin"
  | "Snapchat"
  | "Twitch"
  | "Discord"
  | "Messenger"
  | "Reddit"
  | "GitHub";
 